move activity selectors to seperate folder
closes #272
